Move simpleResponse out to its own file.
