+++ app/user.go
@@ -30,22 +30,6 @@ func makeAdminUserExistsHandler(e *echo.Echo, adb *user.AccountDB) echo.HandlerF
 	}
 }
 
-// simpleResponse is just a handy way to collect errors to return in a
-// JSON payload.
-type simpleResponse struct {
-	Errors []string
-}
-
-// newSimpleResponse creates a simpleResponse with its Errors slice already allocated.
-func newSimpleResponse() *simpleResponse {
-	return &simpleResponse{Errors: []string{}}
-}
-
-// AddError appends an error to a simpleResponses Errors slice.
-func (s *simpleResponse) AddError(err error) {
-	s.Errors = append(s.Errors, err.Error())
-}
-
 // createAccountRequest is a holder for data passed into new user requests.
 type createAccountRequest struct {
 	Forename string `json:"forename"`
